Rename and move the Colors configuration for the Parser.
