+++ src/Parser.Internal.cs
@@ -147,7 +147,7 @@ namespace CommandLine
             // short circuit the request for help!
             if (args.Length == 2 && (args[1] == "/?" || args[1] == "-?"))
             {
-                HelpGenerator.DisplayHelpForCommmand(args[0], arguments.ArgumentGroups[args[0]], Colors.Get());
+                HelpGenerator.DisplayHelpForCommmand(args[0], arguments.ArgumentGroups[args[0]], Configuration.DisplayColors.Get());
 
                 // if we wanted the help, then we successfully parsed it!
                 return true;
