+++ src/Parser.cs
@@ -66,11 +66,11 @@ namespace CommandLine
                 TypeHelpers.ScanTypeForProperties<TOptions>(out TypeArgumentInfo arguments);
 
                 // If we get here, the options type is well defined, so let's display the help.
-                HelpGenerator.DisplayHelp(helpFormat, arguments, Colors.Get());
+                HelpGenerator.DisplayHelp(helpFormat, arguments, Configuration.DisplayColors.Get());
             }
             catch (Exception ex)
             {
-                string errorFormat = $"[{Colors.Get().ErrorColor}!Error]: {{0}} {{1}}";
+                string errorFormat = $"[{Configuration.DisplayColors.Get().ErrorColor}!Error]: {{0}} {{1}}";
                 // If we were asked to display the help and something went wrong, display the error that went wrong.
                 Colorizer.WriteLine(errorFormat, ex.Message, Environment.NewLine);
             }
@@ -97,12 +97,12 @@ namespace CommandLine
                 {
                     if (args[0] == HelpGenerator.RequestShortHelpParameter || args[0] == "/?")
                     {
-                        HelpGenerator.DisplayHelp(HelpFormat.Short, arguments, Colors.Get());
+                        HelpGenerator.DisplayHelp(HelpFormat.Short, arguments, Configuration.DisplayColors.Get());
                         return true;
                     }
                     else if (args[0] == HelpGenerator.RequestLongHelpParameter)
                     {
-                        HelpGenerator.DisplayHelp(HelpFormat.Full, arguments, Colors.Get());
+                        HelpGenerator.DisplayHelp(HelpFormat.Full, arguments, Configuration.DisplayColors.Get());
                         return true;
                     }
                 }
@@ -120,12 +120,12 @@ namespace CommandLine
             catch (Exception innerParserException)
             {
                 ex = new ParserException(innerParserException.Message, innerParserException);
-                if (Options.DisplayErrorMessageOnError)
+                if (Configuration.DisplayErrorMessageOnError)
                 {
-                    string errorFormat = $"[{Colors.Get().ErrorColor}!Error]: {{0}} {{1}}";
+                    string errorFormat = $"[{Configuration.DisplayColors.Get().ErrorColor}!Error]: {{0}} {{1}}";
                     Colorizer.WriteLine(errorFormat, ex.Message, Environment.NewLine);
 
-                    HelpGenerator.DisplayHelp(HelpFormat.Short, arguments, Colors.Get());
+                    HelpGenerator.DisplayHelp(HelpFormat.Short, arguments, Configuration.DisplayColors.Get());
                 }
                 return false;
             }
